user can modify info
