+++ .byebug_history
@@ -1,4 +1,8 @@
 exit
+user_path(user)
+user
+current_path
+exit
 params
 @category
 exit
