+++ app/views/sessions/new.html.erb
@@ -8,7 +8,7 @@
     <%= f.label :password %>
     <%= f.password_field :password %>
 
-    <%= f.submit "Log In", class: "waves-effect waves-light btn light-green darken-4" %>
+    <%= f.submit "Login", class: "waves-effect waves-light btn light-green darken-4" %>
   <% end %>
   <div>
     <p>
